Minor change to player settings

+++ LetsQuiz/Assets/_Game/Scripts/Player/PlayerSettings.cs
@@ -5,11 +5,11 @@ namespace _LetsQuiz
     public class PlayerSettings
     {
         [Header("Sound Effects")]
-        public float soundEffectVolume = -10.0f;
+        public int soundEffectVolume = -10;
         public int soundEffectToggled = 1;
 
         [Header("Background Music")]
-        public float backgroundMusicVolume = -10.0f;
+        public int backgroundMusicVolume = -10;
         public int backgroundMusicToggled = 1;
 
         [Header("Notifications")]
